backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
yfinance==0.2.51
numpy==2.2.2
pandas==2.2.3
scipy==1.15.1
python-dotenv==1.0.1
pydantic==2.10.5
